Gate state is now set correctly when loading the level mid-race.
